pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel>=0.41.0"]
build-backend = "setuptools.build_meta"

[project]
name = "safeskill-agent"
version = "1.0.0"
description = "Runtime-configurable command security enforcement agent for macOS and Linux"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "SafeSkill Team"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: System Administrators",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
]
dependencies = [
    "pyyaml>=6.0.1",
    "watchdog>=4.0.0",
    "jsonschema>=4.21.0",
    "cryptography>=42.0.0",
    "aiohttp>=3.9.0",
    "pydantic>=2.6.0",
    "structlog>=24.1.0",
    "rich>=13.7.0",
    "click>=8.1.0",
    "filelock>=3.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[project.scripts]
safeskill = "safeskill.cli:main"
safeskill-agent = "safeskill.agent:main"

[tool.setuptools.packages.find]
include = ["safeskill*"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "S", "B", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
